Consolidate remaining exec functions into runtime_subprocess.go. Keep chipping away at them
